Feat: 포스트 수정하기
포스트 수정하기 로직 설계
- 포스트 컨텐츠 관련해서 현재 매번 삭제후 새로 추가하는 방식을 사용(수정할때마다 컨테츠 개수가
가변)
- 추후에 로직변경이 필요해보임

+++ src/main/java/com/github/supercodingfinalprojectbackend/repository/PostsSkillStackRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostsSkillStackRepository extends JpaRepository<PostsSkillStack,Integer> {
+public interface PostsSkillStackRepository extends JpaRepository<PostsSkillStack,Long> {
     PostsSkillStack findByPosts(Posts posts);
 }
